requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
autogen-agentchat==0.4.9.3
autogen-core==0.4.9.3
autogen-ext==0.4.9.3
azure-ai-inference==1.0.0b9
azure-core==1.32.0
azure-identity==1.21.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Deprecated==1.2.18
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
jiter==0.9.0
json-schema-to-pydantic==0.2.2
jsonref==1.1.0
mcp==1.6.0
msal==1.32.0
msal-extensions==1.3.1
openai==1.70.0
opentelemetry-api==1.31.1
pathlib==1.0.1
pillow==11.1.0
protobuf==5.29.4
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
PyJWT==2.10.1
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
zipp==3.21.0
